requirements.txt
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
lxml==5.3.0
numpy==2.1.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.11
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.1.2
python-pptx==1.0.2
tabula==1.0.5
typing_extensions==4.12.2
XlsxWriter==3.2.0
